Finished the functionality and layout of the login page
