+++ src/App.jsx
@@ -3,7 +3,8 @@ import { Route, Routes } from "react-router-dom";
 import "./App.scss";
 
 import Home from "./components/Home";
-import Login from "./components/Login";
+import Login from "./components/Login/Login";
+import Profile from "./components/Profile/Profile";
 import Register from "./components/Register/Register";
 
 function App() {
@@ -13,6 +14,7 @@ function App() {
         <Route path="/" element={<Register />} />
         <Route path="/home" element={<Home />} />
         <Route path="/login" element={<Login />} />
+        <Route path="/profile" element={<Profile />} />
       </Routes>
     </div>
   );
