+++ src/components/Profile/Profile.jsx
@@ -2,6 +2,8 @@ import React from "react";
 import { Navigate } from "react-router-dom";
 import { useSelector } from "react-redux";
 
+import styles from "./Profile.module.scss";
+
 const Profile = () => {
   const { user: currentUser } = useSelector((state) => state.auth);
 
@@ -10,10 +12,10 @@ const Profile = () => {
   }
 
   return (
-    <div className="container">
+    <div className={styles.container}>
       <header className="jumbotron">
         <h3>
-          <strong>{currentUser.username}</strong> Profile
+          <strong>{currentUser.user.name}</strong> Profile
         </h3>
       </header>
       <p>
@@ -21,16 +23,19 @@ const Profile = () => {
         {currentUser.accessToken.substr(currentUser.accessToken.length - 20)}
       </p>
       <p>
-        <strong>Id:</strong> {currentUser.id}
+        <strong>Id:</strong> {currentUser.user.id}
       </p>
       <p>
-        <strong>Email:</strong> {currentUser.email}
+        <strong>Email:</strong> {currentUser.user.email}
       </p>
-      <strong>Authorities:</strong>
+      <p>
+        <strong>Age:</strong> {currentUser.user.age}
+      </p>
+      {/* <strong>Authorities:</strong>
       <ul>
         {currentUser.roles &&
           currentUser.roles.map((role, index) => <li key={index}>{role}</li>)}
-      </ul>
+      </ul> */}
     </div>
   );
 };
