+++ src/components/Profile/Profile.module.scss
@@ -1,0 +1,3 @@
+.container {
+  background-color: black;
+}
