+++ src/variables.scss
@@ -1,4 +1,4 @@
-$blak: #161616;
+$black: #161616;
 $gray: rgba(255, 255, 255, 0.38);
 $blue: #157ae1;
 $white: #f9f9f9;
